Fix ServiceEntry WorkloadInstanceHandler for label change in Pods
For a label change in WorkloadInstance, which results in a mismatch with a previously matching ServiceEntry, the whole update event is skipped. Which results in the WorkloadInstance not getting removed from the ServiceEntry.
This fix changes the behaviour for such cases to handle those updates similar to a delete event and clean up the WorkloadInstance from the list.

Fixes [#42921]